refactored task classes into separate files
